feat: created screens folder refactor GoalInput component to screens folder
